### 2025-09-19 #9021 @kthyng
@CaioFerreiraB
* a bit shorter would be better for your paper — you can move extra material to docs as needed
* We have a backlog of submissions so I will add this to our waitlist. Thanks for your patience.
