### 2025-11-20 #9021 @goldingn
Hi @cwitkowitz, @ebezzam would either of you you be interested in reviewing this python package on audio viz for JOSS?
