### 2025-12-10 #9601 @editorialbot
[REVIEW]: Maui: A Python Visualization Framework for Ecoacoustics Data
**Submitting author:** @caioferreirab (<a href="http://orcid.org/0009-0002-9447-8576">Caio Bernardo</a>)
**Repository:** https://github.com/maui-software/maui-software
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.2.7
