### 2026-08-25 #9601 @goldingn
Hi @openjournals/ese-eics this review hasn't heard from the author since September 2025, despite many pings, open issues on the repo, and a direct email to the submitting author. I suggest closing.
